Tanks refactoring 13.07.15 Tanks NETWORK 1.3 with separate Server

+++ src/com/kademika/day11/theory/BAISReader.java
@@ -10,20 +10,20 @@ public class BAISReader {
         byte [] bytes = {1, 34, 45, 24,19, 94, 18, -34, 45, -120, 34};
         ByteArrayInputStream inputStream = new ByteArrayInputStream(bytes);
         printStreamData(inputStream);
-//        byte [] bytes = {1, 34, 45, 24,19, 94, 18, -34, 45, -120, 34};
+//        byte[] bytes = {1, 34, 45, 24, 19, 94, 18, -34, 45, -120, 34};
 //        ByteArrayInputStream inputStream = new ByteArrayInputStream(bytes);
 //        int i;
-//        while((i=inputStream.read()) != -1){
-//            System.out.print(i);
+//        while ((i = inputStream.read()) != -1) {
+//            System.out.print((byte)i);
 //            System.out.print(" ");
 //        }
-//
+//    }
     }
 
     public static void printStreamData(ByteArrayInputStream byteArrayInputStream){
                 int i;
         while((i=byteArrayInputStream.read()) != -1){
-            System.out.print(i);
+            System.out.print((byte)i);
             System.out.print(" ");
         }
     }
